new algos

+++ src/CharCount.java
@@ -5,16 +5,16 @@ public class CharCount {
 	  System.out.println(repeatedString("aba", 10));
 	}
 
-	static int repeatedString(String s, int n) {
-		int totalNumberOfPossibleSubString = n / s.length();
-		int remainingPartOfString = n % s.length();
+  static long repeatedString(String s, long n) {
+    long totalNumberOfPossibleSubString = n / s.length();
+    long remainingPartOfString = n % s.length();
 
 		if (!s.contains("a"))
 			return 0;
 		return s.length() > n ? count(s, n) : totalNumberOfPossibleSubString * count(s, s.length()) + count(s, remainingPartOfString);
 	}
 
-	static int count(String s, int e) {
+  static long count(String s, long e) {
 		int a = 0;
 		for (int i = 0; i < e; i++) {
 			if (s.charAt(i) == 'a')
